+++ src/tooling/piral-cli/src/types/internal.ts
@@ -8,6 +8,7 @@ export interface PiletDefinition {
     string,
     {
       selected?: boolean;
+      port?: number;
     }
   >;
 }